Compare two data frames

Compare two data frames

Usage

compare(
  table_a,
  table_b,
  by,
  allow_both_NA = TRUE,
  coerce = TRUE,
  use_duckplyr = FALSE
)

value_diffs(comparison, column)

all_value_diffs(comparison)

Arguments

table_a

A data frame

table_b

A data frame

by

tidy-select. Selection of columns to use when matching rows between .data_a and .data_b. Both data frames must be unique on by.

allow_both_NA

Logical. If TRUE a missing value in both data frames is considered as equal

coerce

Logical. If FALSE only columns with the same class are compared.

use_duckplyr

Logical. Run as_duckplyr_df() on input tables before performing comparison. This is useful when the tables are large because it makes the comparison faster. If TRUE, the outputs will also be duckplyr_df objects.

comparison

The output of a versus::compare() call

column

tidy-select. A single column

Value

compare()

A list of data frames having the following elements:

tables

A data frame with one row per input table showing the number of rows and columns in each.

by

A data frame with one row per by column showing the class of the column in each of the input tables.

summ

A data frame with one row per column common to table_a and table_b and columns "n_diffs" showing the number of values which are different between the two tables, "class_a"/"class_b" the class of the column in each table, and "value_diffs" a (nested) data frame showing the the values in each table which are unequal, and the by columns

unmatched_cols

A data frame with one row per column which is in one input table but not the other and columns "table": which table the column appears in, "column": the name of the column, and "class": the class of the column.

unmatched_rows

A data frame which, for each row present in one input table but not the other, contains the column "table" showing which table the row appears in and the by columns for that row.

value_diffs()

A data frame with one row for each element of col found to be unequal between the input tables ( table_a and table_b from the original compare() call) The output table has columns "val_a"/"val_b": the value of col in the input tables, and the by columns for the identified rows in the input tables.

all_value_diffs()

A data frame of the value_diffs() output for all columns having at least one value difference, combined row-wise into a single table. To facilitate this combination into a single table, the "val_a" and "val_b" columns are coerced to character.