Show rows which have differing values in a comparison

Show rows which have differing values in a comparison

Usage

slice_diffs(table, comparison, column = everything())

slice_diffs_both(table_a, table_b, comparison, column = everything())

Arguments

table

A data frame. One of table_a or table_b used to create comparison

comparison

The output of compare()

column

<tidy-select>. A row will be in the output if the comparison shows differing values for any columns matching this argument

table_a

The data frametable_a used to create comparison

table_b

The data frametable_b used to create comparison

Value

slice_diffs()

The input table is filtered to the rows for which comparison shows differing values for one of the columns in the supplied column argument

slice_diffs_both()

The output of slice_diffs() for both input tables with the rows interleaved and a column table indicating which table the row is from. The output contains only columns present in both tables.

Examples

comp <- compare(example_df_a, example_df_b, by = car)
example_df_a |> slice_diffs(comp, c(disp, mpg))
#> # A tibble: 4 × 9
#>   car              mpg   cyl  disp    hp  drat    wt    vs    am
#>   <chr>          <dbl> <int> <dbl> <int> <dbl> <dbl> <int> <int>
#> 1 Duster 360      14.3     8  360    245  3.21  3.57     0     0
#> 2 Datsun 710      22.8    NA  109     93  3.85  2.32     1     1
#> 3 Merc 240D       24.4     4  147.    62  3.69  3.19     1     0
#> 4 Hornet 4 Drive  21.4     6  259    110  3.08  3.22     1     0
slice_diffs_both(example_df_a, example_df_b, comp, column = disp)
#> # A tibble: 4 × 9
#>   table car              mpg   cyl  disp    hp  drat    wt    vs
#>   <chr> <chr>          <dbl> <int> <dbl> <int> <dbl> <dbl> <int>
#> 1 a     Datsun 710      22.8    NA   109    93  3.85  2.32     1
#> 2 b     Datsun 710      22.8    NA   108    93  3.85  2.32     1
#> 3 a     Hornet 4 Drive  21.4     6   259   110  3.08  3.22     1
#> 4 b     Hornet 4 Drive  21.4     6   258   110  3.08  3.22     1